working cubemap on opengl2

+++ learnopengl/cubemapemcc/libs/resources/shaders/skybox.fs
@@ -1,5 +1,5 @@
 #version 300 es
-precision highp float;
+precision mediump float;
 
 in vec3 TexCoords;
 out vec4 FragColor;
@@ -7,6 +7,6 @@ out vec4 FragColor;
 uniform samplerCube skybox;
 
 void main()
-{    
+{
     FragColor = texture(skybox, TexCoords);
 }
